Fix null pointers, authorization and other review comments

+++ OmgvaPOS/src/ItemManagement/Services/ItemService.cs
@@ -1,4 +1,6 @@
-﻿using OmgvaPOS.TaxManagement.Repository;
+﻿using OmgvaPOS.DiscountManagement.Service;
+using OmgvaPOS.Exceptions;
+using OmgvaPOS.TaxManagement.Repository;
 using OmgvaPOS.ItemManagement.Repositories;
 using OmgvaPOS.ItemVariationManagement.Repositories;
 using OmgvaPOS.ItemManagement.Models;
@@ -7,6 +9,7 @@ using OmgvaPOS.ItemManagement.Mappers;
 using OmgvaPOS.ItemManagement.Validator;
 using OmgvaPOS.TaxManagement.Models;
 using OmgvaPOS.TaxManagement.Mappers;
+using OmgvaPOS.UserManagement.Service;
 
 namespace OmgvaPOS.ItemManagement.Services
 {
@@ -15,6 +18,8 @@ namespace OmgvaPOS.ItemManagement.Services
         IItemVariationRepository itemVariationRepository,
         ITaxRepository taxRepository,
         ITaxItemRepository taxItemRepository,
+        IUserService userService,
+        DiscountValidatorService discountValidatorService,
         ILogger<ItemService> logger
         ) : IItemService
     {
@@ -22,6 +27,8 @@ namespace OmgvaPOS.ItemManagement.Services
         private readonly ITaxItemRepository _taxItemRepository = taxItemRepository;
         private readonly IItemRepository _itemRepository = itemRepository;
         private readonly IItemVariationRepository _itemVariationRepository = itemVariationRepository;
+        private readonly IUserService _userService = userService;
+        private readonly DiscountValidatorService _discountValidatorService = discountValidatorService;
         private readonly ILogger<ItemService> _logger = logger;
 
         //TODO: Think how you will add taxes, discounts
@@ -30,28 +37,52 @@ namespace OmgvaPOS.ItemManagement.Services
             var itemDTOs = items.Select(i => i.ToItemDTO()).ToList();
             return itemDTOs;
         }
-        public ItemDTO GetItem(long id) {
+        public ItemDTO? GetItem(long id) {
             var item = _itemRepository.GetItem(id);
-            return item.ToItemDTO();
+            return item?.ToItemDTO();
         }
 
         public Item GetItemNoException(long id) {
             return _itemRepository.GetItem(id);
         }
+
+        private Item GetItemOrThrow(long itemId)
+        {
+            var item = _itemRepository.GetItem(itemId);
+            if (item == null)
+                throw new NotFoundException("Item not found");
+
+            return item;
+        }
+
+        public long GetItemBusinessId(long itemId)
+        {
+            return GetItemOrThrow(itemId).BusinessId;
+        }
+        
         public ItemDTO CreateItem(CreateItemRequest createItemRequest, long businessId) {
-            //TODO: Think about discounts
             createItemRequest.Currency = createItemRequest.Currency.ToUpper();
             ItemValidator.ValidateCreateItemRequest(createItemRequest);
+            
+            _userService.ValidateUserBelongsToBusiness(createItemRequest.UserId, businessId);
+            _discountValidatorService.ValidateDiscountBelongsToBusiness(createItemRequest.DiscountId, businessId);
             
             var newItemData = createItemRequest.ToItem(businessId);
             var newItem = _itemRepository.CreateItem(newItemData);
             
             return newItem.ToItemDTO();
         }
-        public ItemDTO? UpdateItem(ItemDTO updateItemRequest) {
-            var item = _itemRepository.GetItem((long)updateItemRequest.Id); //TODO: potential error here. Though unlikely as endpoint should make sure of id existance
-            item = updateItemRequest.FromUpdateRequestToItem(item);
-            return UpdateItem(item).ToItemDTO();
+        
+        public ItemDTO UpdateItem(ItemDTO updateItemRequest, long itemId) {
+            var currentItem = _itemRepository.GetItem(itemId);
+            if (currentItem == null)
+                throw new NotFoundException();
+            
+            _userService.ValidateUserBelongsToBusiness(updateItemRequest.UserId, currentItem.BusinessId);
+            _discountValidatorService.ValidateDiscountBelongsToBusiness(updateItemRequest.DiscountId, currentItem.BusinessId);
+            
+            currentItem = updateItemRequest.ToItem(currentItem);
+            return UpdateItem(currentItem).ToItemDTO();
         }
 
         private Item UpdateItem(Item item) {
@@ -76,7 +107,11 @@ namespace OmgvaPOS.ItemManagement.Services
                 _itemVariationRepository.DuplicateItemVariations(itemVariations, recreatedItem.Id);
             }
         }
-        public void DeleteItem(long id) {
+        public void DeleteItem(long id)
+        {
+            if (_itemRepository.GetItem(id) == null)
+                throw new NotFoundException();
+            
             var itemVariations = _itemVariationRepository.GetItemVariationQueriable().Where(v => v.ItemId == id);
             _itemVariationRepository.DeleteItemVariations(itemVariations.Select(v => v.Id));
             _itemRepository.DeleteItem(id);
@@ -89,12 +124,15 @@ namespace OmgvaPOS.ItemManagement.Services
             var taxItems = _taxItemRepository.GetAllTaxItemQueriable().Where(c => c.ItemId == id);
             return _taxRepository.GetAllTaxes()
                 .Where(t => taxItems.Select(c => c.TaxId).Contains(t.Id))
-                .Select(t => TaxMapper.ToDTO(t)).ToList();
+                .Select(TaxMapper.ToDTO).ToList();
         }
-        public ItemDTO ChangeItemTaxes(ChangeItemTaxesRequest changeItemTaxesRequest, long itemId) {
-            // TODO: REWRITE for better clarity. 
-            var newItem = UpdateItem(_itemRepository.GetItem(itemId)); //TODO: Potential error here
-
+        public ItemDTO ChangeItemTaxes(ChangeItemTaxesRequest changeItemTaxesRequest, long itemId)
+        {
+            var currentItem = _itemRepository.GetItem(itemId);
+            if (currentItem == null)
+                throw new NotFoundException();
+            
+            var newItem = UpdateItem(currentItem);
             var taxItems = _taxItemRepository.GetAllTaxItemQueriable();
             
             var taxItemIdsToRemove = taxItems
